app/requirements.txt
flask==2.3.3
flask_restx==1.2.0
pymysql==1.1.0
mysql-connector==2.2.9
python-dotenv==1.0.0